requirements.txt
albumentations==1.3.0
antlr4-python3-runtime==4.9.3
anyconfig==0.10.1
anyio==3.6.2
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astroid==2.12.13
asttokens==2.2.1
attrs==22.1.0
autopep8==2.0.1
backcall==0.2.0
beautifulsoup4==4.11.1
binaryornot==0.4.4
bleach==5.0.1
build==0.10.0
cachetools==5.3.0
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_477u68wvzm/croot/certifi_1671487773341/work/certifi
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
click-config-file==0.6.0
click-option-group==0.5.5
cloudpickle==2.2.0
comm==0.1.2
configobj==5.0.8
contourpy==1.0.6
cookiecutter==2.1.1
cycler==0.11.0
Cython==0.29.34
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
dice-ml==0.9
dill==0.3.6
docker-pycreds==0.4.0
dynaconf==3.1.11
einops==0.6.0
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.9.0
fonttools==4.38.0
fqdn==1.5.1
FrEIA==0.2
fsspec==2023.1.0
gitdb==4.0.10
GitPython==3.1.31
h5py==3.7.0
hdbscan==0.8.29
huggingface-hub==0.12.1
idna==3.4
ignite==1.1.0
imageio==2.23.0
importlib-metadata==6.0.0
importlib-resources==5.12.0
install==1.3.5
ipykernel==6.19.3
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.3
isoduration==20.11.0
isort==5.11.3
jedi==0.18.2
Jinja2==3.1.2
jinja2-time==0.2.0
jinja2schema==0.1.4
jmespath==0.10.0
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.5.0
jupyter_client==7.4.8
jupyter_core==5.1.0
jupyter_server==2.0.1
jupyter_server_terminals==0.4.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.4
kedro==0.18.5
kiwisolver==1.4.4
kmeans-pytorch==0.3
lazy-object-proxy==1.8.0
llvmlite==0.39.1
Logbook==1.5.3
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==2.0.4
more-itertools==9.1.0
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.6
networkx==2.8.8
notebook==6.5.2
notebook_shim==0.2.2
numba==0.56.4
numpy==1.23.5
omegaconf==2.3.0
opencv-python-headless==4.7.0.68
packaging==22.0
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pip-tools==6.12.2
platformdirs==2.6.0
pluggy==1.0.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==4.22.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
PyClone==2020.9b2
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.14.0
pylint==2.15.9
pyparsing==3.0.9
pyproject_hooks==1.0.0
pyrsistent==0.19.2
python-dateutil==2.8.2
python-json-logger==2.0.4
python-slugify==8.0.1
pytoolconfig==1.2.5
pytorch-ignite==0.4.11
pytorch-pretrained-vit==0.0.7
pytz==2022.7
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
qudida==0.0.4
requests==2.28.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.1
rope==1.7.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-image==0.19.3
scikit-learn==1.2.0
scipy==1.9.3
seaborn==0.12.2
Send2Trash==1.8.0
sentry-sdk==1.16.0
setproctitle==1.3.2
shap==0.41.0
six==1.16.0
sklearn==0.0.post1
slicer==0.0.7
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
swin-transformer-pytorch==0.4.1
tenacity==8.2.2
terminado==0.17.1
text-unidecode==1.3
threadpoolctl==3.1.0
tifffile==2022.10.10
timm==0.6.13
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
toposort==1.10
torch==1.13.1
torchmetrics==0.11.3
torchsummary==1.5.1
torchsummaryX==1.3.0
torchtyping==0.1.4
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
traitlets==5.8.0
typeguard==2.13.3
typing_extensions==4.4.0
uri-template==1.2.0
urllib3==1.26.13
vit-pytorch==0.40.2
wandb==0.14.0
wandb-utils==0.1.2
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
widgetsnbextension==4.0.4
wrapt==1.14.1
xgboost==1.7.3
yacs==0.1.8
zipp==3.15.0
